Add MigrationAccount
MigrationAccount is a data structure representing
the external Klaytn account that the user wants to transfer to KAS.

User can specify address, nonce, and MigrationAccountKey.
MigrationAccount must contains address and MigrationAccountKey for migration.

MigrationAccountKey is a data structure representing
variable types of private key (RoleBased, Multisig, Single).
This is used to sign FeeDelegatedAccountUpdateTransaction.