Fix wrong exception handling at account migration test cases
1. Change test case names
Below methods does not throw exceptions, but get 200 Http status code as response from KAS. So `ThrowException` does not fit in this case.
* migrateSingleKeyAccount_ThrowException_AllFailed -> migrateSingleKeyAccount_AllFailed
* migrateMultipleSingleKeyAccounts_throwException_PartiallyFailed -> migrateMultipleSingleKeyAccounts_PartiallyFailed

2. Add try and catch to test cases which should success
Put a try & catch statement in the test cases that should succeed.
This is because unintended exceptions must be handled.

3. Handle Exception using expectedException
This is more readable and maintainable methods to handle expected exceptions.