Update test case and exception message
If SDK user who did not initialize node api, rpc call try to send request to Caver.DEFAULT_URL(= http://localhost:8551)

Since it is unlikely that the KAS service will be run on the local host and tested, this test case will be kept as it is.

When a user directly launches the KAS service on the local host, it is when the user is the KAS development team.
Most users will not fall into this case, so we will keep this test case as it is.